22nd December 2025
Today, i have met and created these

- Creating the model for the floor

- Main Menu Screen (play button and exit)

- Create the Player’s Model (Ball)

Which can all be checked on the Trello Board. alongside my creation of all of the trial elements. Please also view the Recording that will be in the videos folder title 22/12/2025. It will have a recorded speed up version of everything i have created.