29.12.2025
Added new features like, the Coins to the level scene, coin counter variable into the Globals so it can be called and use a Area3D to detect when a coin is touched by the player it will then add +1 to the counter of it. as well as animation of the coin so it now rotates.